fix: move resolveAsaForKey to agent-memories to avoid mock.module leak
The tool-handlers test uses mock.module on arc69-store which globally
replaces resolveAsaForKey with a stub returning null, causing the
arc69-store test to fail when run in the same process. Moving the
pure DB function to agent-memories (where it logically belongs) and
re-exporting from arc69-store fixes the leak while preserving the API.
