+++ server/memory/arc69-store.ts
@@ -347,16 +347,5 @@ export async function listMemoryAsas(
     return memories;
 }
 
-/**
- * Look up the ASA ID for a given memory key from the local DB mapping.
- */
-export function resolveAsaForKey(
-    db: Database,
-    agentId: string,
-    key: string,
-): number | null {
-    const row = db.query(
-        'SELECT asa_id FROM agent_memories WHERE agent_id = ? AND key = ? AND asa_id IS NOT NULL'
-    ).get(agentId, key) as { asa_id: number } | null;
-    return row?.asa_id ?? null;
-}
+// Re-export from agent-memories for backward compatibility (dynamic imports use this path)
+export { resolveAsaForKey } from '../db/agent-memories';
